requirements.txt
gitdb==4.0.11
GitPython==3.1.40
PyYAML==6.0.1
smmap==5.0.1
